requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
fastapi==0.115.12
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
openai==1.0.0
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2